backend/requirements-dev.txt
-r requirements.txt
fastapi[testclient]
sqlalchemy-utils==0.41.1
httpx==0.26.0
ruff==0.3.4
black==23.12.1
pytest-cov==4.1.0
